♻️ refactor(mq-markdown): extract TableLayout into markdown/table_layout.rs
Move table width/alignment computation and cell padding logic out of
markdown.rs into a dedicated submodule, mirroring the existing
node.rs/node/ and html_to_markdown.rs/html_to_markdown/ split.
